requirements_jittor.txt
jittor
numpy
opencv-python
pyyaml
easydict
tensorboard
Pillow
